Fix 32-bit builds

+++ storm/Thread.cpp
@@ -8,12 +8,12 @@
 #include <pthread.h>
 #endif
 
-uintptr_t SGetCurrentThreadId() {
+long unsigned int SGetCurrentThreadId() {
 #if defined(WHOA_SYSTEM_WIN)
     return GetCurrentThreadId();
 #endif
 
 #if defined(WHOA_SYSTEM_MAC) || defined(WHOA_SYSTEM_LINUX)
-    return reinterpret_cast</*long unsigned int*/ intptr_t>(pthread_self());
+    return reinterpret_cast<long unsigned int>(pthread_self());
 #endif
 }
